common/requirements.txt
Appium-Python-Client==0.48
APScheduler==3.6.1
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.1
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.3.1
cssselect==1.1.0
Django==2.2.6
django-cors-headers==3.1.1
docopt==0.6.2
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Login==0.4.1
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
gerapy==0.8.5
h11==0.7.0
h2==3.1.1
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
kaitaistruct==0.8
ldap3==2.5.2
logparser==0.8.2
lxml==4.4.1
MarkupSafe==1.1.1
mitmproxy==4.0.4
Naked==0.1.31
numpy==1.17.2
opencv-python==4.1.1.26
openpyxl==3.0.0
pandas==0.25.1
parsel==1.5.2
passlib==1.7.1
pexpect==4.7.0
Pillow==6.1.0
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodome==3.9.0
PyDispatcher==2.0.5
pydivert==2.1.0
PyExecJS==1.5.1
PyHamcrest==1.9.0
pymongo==3.9.0
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
python-dateutil==2.8.0
python-scrapyd-api==2.1.2
pytz==2019.2
PyYAML==5.1.2
queuelib==1.5.0
redis==3.3.8
requests==2.22.0
ruamel.yaml==0.15.100
scipy==1.3.1
Scrapy==1.7.3
scrapy-redis==0.6.8
scrapy-splash==0.7.2
scrapyd==1.2.1
scrapyd-client==1.1.0
scrapydweb==1.4.0
selenium==3.141.0
service-identity==18.1.0
shellescape==3.4.1
six==1.12.0
sortedcontainers==2.0.5
soupsieve==1.9.4
SQLAlchemy==1.3.8
sqlparse==0.3.0
tornado==5.1.1
Twisted==19.7.0
tzlocal==2.0.0
urllib3==1.25.6
urwid==2.0.1
w3lib==1.21.0
Werkzeug==0.16.0
wsproto==0.11.0
xlrd==1.2.0
zope.interface==4.6.0
